pyproject.toml
[tool.poetry]
name = "taipei-day-trip"
version = "0.1.0"
description = ""
authors = ["ChengTze-Wu <yaoop3050777@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
certifi = "2021.10.8"
charset-normalizer = "2.0.12"
click = "8.0.4"
flask = "2.0.3"
gunicorn = "20.1.0"
idna = "3.3"
itsdangerous = "2.1.0"
jinja2 = "3.0.3"
markupsafe = "2.1.0"
mysql-connector-python = "8.0.28"
protobuf = "3.19.4"
pyjwt = "2.3.0"
python-dotenv = "0.19.2"
requests = "2.27.1"
urllib3 = "1.26.9"
werkzeug = "2.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
